Configuration of descriptors is now working
It is possible to define an extension with a descriptor and have the configurations for the descriptor saved, and used by the plugin. The extension points are not yet used in the plugin, but should be done through an Action. There is currently a problem with the parametisating of the SCMInterface class which should be solved, as the AbstractSCMInterface currently does not implement this